+++ README.md
@@ -19,5 +19,5 @@
 
 ### 3단계
 - [X] JSON View 구현
-- [] UserController 컨트롤러가 json 형태로 응답을 반환한다.
-- [] 레거시 컨트롤러를 제거해도 서버가 정상 동작한다.
+- [X] UserController 컨트롤러가 json 형태로 응답을 반환한다.
+- [X] 레거시 컨트롤러를 제거해도 서버가 정상 동작한다.
