+++ mvc/src/main/java/nextstep/mvc/AnnotationHandlerAdapter.java
@@ -15,6 +15,6 @@ public class AnnotationHandlerAdapter implements HandlerAdapter {
     @Override
     public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response,
                                final Object handler) throws Exception {
-        return ((HandlerExecution) handler).handle(request, response);
+        return (ModelAndView) ((HandlerExecution) handler).handle(request, response);
     }
 }
