+++ mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerExecution.java
@@ -3,6 +3,7 @@ package nextstep.mvc.controller.tobe;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.lang.reflect.Method;
+import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 
 public class HandlerExecution {
@@ -15,7 +16,11 @@ public class HandlerExecution {
         this.method = method;
     }
 
-    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response) throws Exception {
-        return (ModelAndView) method.invoke(handler, request, response);
+    public Object handle(final HttpServletRequest request, final HttpServletResponse response) throws Exception {
+        final Object invoke = method.invoke(handler, request, response);
+        if (invoke instanceof ModelAndView) {
+            return method.invoke(handler, request, response);
+        }
+        return new ModelAndView(new JspView((String) method.invoke(handler, request, response)));
     }
 }
